requirements.txt
catboost==1.2.0
geopy==2.4.1
kafka-python==2.0.2
psycopg2-binary==2.9.3
flask==2.0.1
numpy==1.23.5
pandas==1.4.3
scikit-learn==1.0.2
matplotlib==3.5.1
seaborn==0.11.2